fixed merge comments
